+++ src/test/java/com/inq/wishhair/wesharewishhair/review/service/ReviewFindServiceTest.java
@@ -1,0 +1,50 @@
+package com.inq.wishhair.wesharewishhair.review.service;
+
+import com.inq.wishhair.wesharewishhair.fixture.HairStyleFixture;
+import com.inq.wishhair.wesharewishhair.fixture.ReviewFixture;
+import com.inq.wishhair.wesharewishhair.fixture.UserFixture;
+import com.inq.wishhair.wesharewishhair.global.base.ServiceTest;
+import com.inq.wishhair.wesharewishhair.hairstyle.domain.HairStyle;
+import com.inq.wishhair.wesharewishhair.review.domain.Review;
+import com.inq.wishhair.wesharewishhair.user.domain.User;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+@DisplayName("ReviewFindServiceTest - SpringBootTest")
+public class ReviewFindServiceTest extends ServiceTest {
+
+    @Autowired
+    private ReviewFindService reviewFindService;
+
+    private Review review;
+    private User user;
+    private HairStyle hairStyle;
+
+    @BeforeEach
+    void setUp() {
+        user = userRepository.save(UserFixture.B.toEntity());
+        hairStyle = hairStyleRepository.save(HairStyleFixture.A.toEntity());
+        review =  reviewRepository.save(ReviewFixture.A.toEntity(user, hairStyle));
+    }
+
+    @Test
+    @DisplayName("아이디로 리뷰를 조회한다")
+    void findById() {
+        //when
+        Review result = reviewFindService.findById(review.getId());
+
+        //then
+        assertAll(
+                () -> assertThat(result.getUser()).isEqualTo(user),
+                () -> assertThat(result.getHairStyle()).isEqualTo(hairStyle),
+                () -> assertThat(result.getScore()).isEqualTo(review.getScore()),
+                () -> assertThat(result.getContents()).isEqualTo(review.getContents())
+        );
+    }
+}
